event_booking_backend/requirements.txt
Django==5.2.8
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
gunicorn==22.0.0
python-dotenv==1.2.1
langchain==0.3.7
langchain-openai==0.2.7
openai==1.55.3
upstash-vector==0.8.0
httpx==0.28.1
